Fixed bugs #2
