+++ src/features/StreetcodePage/ArtGalleryBlock/ArtGalleryBlock.styles.scss
