+++ src/features/StreetcodePage/Streetcode.component.tsx
@@ -10,7 +10,7 @@ import RelatedFiguresComponent from '@streetcode/RelatedFiguresBlock/RelatedFigu
 import SourcesComponent from '@streetcode/SourcesBlock/Sources.component';
 import TextComponent from '@streetcode/TextBlock/TextBlock.component';
 import TickerComponent from '@streetcode/TickerBlock/Ticker.component';
-import ArtGallery from './ArtGallery/ArtGallery.component';
+import ArtGalleryBlockComponent from '@streetcode/ArtGalleryBlock/ArtGalleryBlock.component';
 
 const StreetcodeContent = () => (
     <div className="streetcodeContainer">
@@ -18,7 +18,7 @@ const StreetcodeContent = () => (
         <MainBlock />
         <TextComponent />
         <InterestingFactsComponent />
-        <ArtGallery />
+        <ArtGalleryBlockComponent />
         <RelatedFiguresComponent />
         <SourcesComponent />
         <QRComponent />
